wifi-extractor/requirements.txt
colorama
prettytable
python-dotenv
subprocess 
re 
smtplib
os
MIMEText
MIMEMultipart
